requirements.txt
streamlit
python-dotenv
google-generativeai
gtts
Pillow
streamlit-geolocation
qdrant-client